6월/250616/requirements.txt
# 라이브러리 의존성 명시
# - 의존성 (dependencies)
#   이 애플리케이션 구동에 꼭 필요하다.

django-environ

# 모든 python 애플리케이션 저장소에는 반드시 있는 파일
# 해당 애플맄이션 구동에 필요한 모든 라이브러리들을 1줄에 하나씩 명시
# 한방 설치: pip install -r requirements.txt